+++ src/overlays/actors/ovl_En_Rat/z_en_rat.c
@@ -324,9 +324,9 @@ void EnRat_ChooseDirection(EnRat* this) {
                 angle -= 0x8000;
             }
 
-            angle += (s16)Rand_CenteredFloat(0x800);
+            angle += (s16)(s32)Rand_CenteredFloat(0x800);
         } else {
-            angle = (Rand_ZeroOne() < 0.1f) ? (s16)Rand_CenteredFloat(0x800) : 0;
+            angle = (Rand_ZeroOne() < 0.1f) ? (s16)(s32)Rand_CenteredFloat(0x800) : 0;
         }
     }
